requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.26.2
Pillow==10.1.0
scikit-learn==1.3.2
scipy==1.11.4
threadpoolctl==3.2.0
Werkzeug==3.0.1
